+++ webpack.prod.config.js
@@ -17,7 +17,13 @@ var config = {
     loaders: [{
       test: /\.(js|jsx)/,
       loader: 'babel',
-    }],
+    },
+    {
+      test: /\.css$/,
+      loader: 'style!css',
+      include: __dirname
+    }
+    ],
   },
   plugins: [],
   resolve: {