+++ HISTORY.md
@@ -1,8 +1,12 @@
+## 0.2.0 (11/26/2015)
+- issue #5: Initial expanded paths
+- CSS and JS bundled together in the build by default
+
 ## 0.1.6 (10/26/2015)
-- Upgrade to React 0.14 
+- Upgrade to React 0.14
 
 ## 0.1.5 (09/10/2015)
-- Display dates in string form 
+- Display dates in string form
 - Resolve issue #6 "undefined" is rendered as Object
 
 ## 0.1.0 (08/06/2015)